+++ libraries/slip10/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.2.1](https://github.com/ExodusMovement/exodus-hydra/compare/@exodus/slip10@2.2.0...@exodus/slip10@2.2.1) (2024-11-03)
+
+**Note:** Version bump only for package @exodus/slip10
+
 ## [2.2.0](https://github.com/ExodusMovement/exodus-hydra/compare/@exodus/slip10@2.1.0...@exodus/slip10@2.2.0) (2024-11-03)
 
 ### Features